refactor: extract bottom buttons to standalone component
